Refactor organization
The project organization was refactored to include all packages in
the main stylecodes package. This was done to avoid importing many
packages to get full use of the lib which would be cumbersome.

Also, StripANSI() function has been exported as this is a useful
function to have when dealing with ANSI escape codes. Getting the
len of a string with escape codes in it will not return an expected
result, and this function helps with that.