Use constants for escape codes
This commit uses constants for the escape codes as it alligns with
idomatic Go. Previously, the escape codes were contained within
structs for sake of organization, but structs are not typically used
for this purpose in Go. For example, in the Go http package, status
codes are not kept in a big 'Status' struct with fields like 'Ok'.
They are constants following a naming convention 'Status_' like
'StatusOk'. The new constants follow a similar naming convention,
'Color_', 'BgColor_', or 'Reset_'. Modifiers just use the name of the
modifier because their names are unique and don't need to be differentiated.